fix: moved ssh changes back into the original file
